speech-gateway/requirements.txt
numpy<2.0
wyoming-openwakeword
asyncio
requests

